型 ReceiptNumber  のベースを変更

+++ Core/Include/Account/Common/AccountsTypes.h
@@ -172,7 +172,7 @@ DECLARE_STRICT_TYPE(int,        PurchaseNumber);
 **    レシートインデックス型。
 **/
 
-DECLARE_STRICT_TYPE(int,        ReceiptNumber);
+DECLARE_STRICT_TYPE(size_t,     ReceiptNumber);
 
 //----------------------------------------------------------------
 /**
